Basic OpMode class & @UsesKotlin
The OpMode class overrides the init() method from it's superclass,
and prevents subclasses from overriding it again; however, its
functionality has been replaced with the new abstract method
`initRobot()`, which is automatically called at the end of the
init() method.

Because reflections in Kotlin is slightly different than it is
in Java, we need OpModes that are written in Kotlin to have the
@UsesKotlin annotation on the class so the API will be able to
adapt and perform the correct tasks.