Added @Hardware annotation
@Hardware is used to annotation properties or fields that represent
a piece of hardware (i.e. a DcMotor or Servo, etc.) that were
configured by the device mapping.

By using this annotation, when the OpMode is initialized, the value
of the Hardware device will be grabbed from the hardware map and
set as the value of the property or field.