Reworking the OpMode code
Now, an OpMode is only created (by the OpModeWrapper class) after all the hardware
has been found and data initialized. When the OpModeWrapper is init()'d, it updates
the current Hardware information (in the 'Hacks' file) with the hardware information
that it has, then creates the OpMode (so that now it can initialize with all the
hardware completely initialized and taken care of) and runs it's init() method.

Every method in the OpModeWrapper is simply a cover method for the actual OpMode's.