Super hacky workaround to get DeviceMapping type
For some reason, the good ol' hack I wanted to use (just straight, getGenericClass( deviceMap ))
wouldn't work because it's genericSuperType couldn't be casted to a ParameterizedType; however,
there is a backing field for the DeviceMapping, a private HashMap< String, * > named "a".

By using getDelcaredField( "a" ), we can access this field; however we first must force it
to be accessible, which is, thankfully, only one line long, then we can grab the instance of
it using our DeviceMapping instance, and the map's values collection's genericSuperType is,
in fact, ParameterizedType, so getGenericClass( map.values ) works.

This is shit.

On one hand, I love how _bad_ this is, but, on the other, I hate myself for actually
deciding that I needed to write this.

Oh well :^). It works!