Added long-press action to the robot icon
In another hack like the one to get the current activity, I managed to get to the robot icon
(in the top left corner of the robot controller activity) via nothing but reflections and
trust. :^) (*no* dependencies on the FtcRobotController module either, so it had to be
completely through reflections, which is awful)

While the program instantly crashes whenever it switches to the ConfigActivity after long
pressing on the robot icon at the moment, this will soon be fixed.