Removed Int as a configurable data type
Only long is supported so that there's no confusion the in the framework
about what the max/min data types should be, or how to handle the data type.
Also, there is no real reason not to use longs, it's safer for overflows and
memory shouldn't be an issue in these applications.