Fixed getDevice< T >
I didn't fully check how the method was running previously, so I wasn't
able to tell that it would never pull the correct device mapping for the
given type due to how the getGenericType() method ran. :/

This new method isn't super efficient, but hopefully it won't be an issue
for anyone who uses it.