Fixed class-determination and logic flaw
Once again, I forgot that `when` doesn't really work too well
when comparing classes, so I had to switch to an if/else if/else
block.

also, the logic for (long == null) xor (double == null) was the exact
opposite of what I wanted. That should be true, so I added a NOT
at the beginning