Support for 2-param constructors in HardwareExts
HardwareExtension constructors can have a second parameter if it's a String.
The value passed as a string will be the name of the device in the hardware
map, for logging and what not.

In the event that two valid constructors exist, one with just the device,
and the other with the device and its name, the latter will always be chosen
as it is more specific than the former. Also, that's just how I set it up
and I won't change it unless it's actually a problem.