Updated libraries, fixed issues
Actually fixed the stupid deprecated method warnings

Updated the libraries in ftc-ext to match those in FtcRobotController so I won't
get weird things that technically build but the IDE says wouldn't. Also, y'know,
it'd be nice if the app ran.

Combined what's left of Hacks.kt with R.kt to make Android.kt