New Configuration System design
This design will better separate the individual sections of the configuration
system from each other in a manner that will also help manner the fragments
that are used to edit them.

As is visible in the diagram, now each fragment used in the Configuration Editor
will have a counterpart in the actual design of the configuration system.
This will hopefully make the process much cleared than it is now.